Turned location on

+++ UWP-Project/Model/WeatherAttributes.cs
@@ -24,7 +24,7 @@ namespace UWP_Project.Model
     {
         public double temp { get; set; }
         public double pressure { get; set; }
-        public double humidity { get; set; }
+        public int humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
         public double sea_level { get; set; }
